Pass controller values to Clients::Page::* components, update activity base class
