+++ ui/app/components/clients/activity.ts
@@ -10,15 +10,18 @@ import Component from '@glimmer/component';
 import { isSameMonth } from 'date-fns';
 import { parseAPITimestamp } from 'core/utils/date-formatters';
 import { calculateAverage } from 'vault/utils/chart-helpers';
-import { filterVersionHistory, hasMountsKey, hasNamespacesKey } from 'core/utils/client-count-utils';
+import { filterVersionHistory } from 'core/utils/client-count-utils';
 
 import type ClientsActivityModel from 'vault/models/clients/activity';
 import type ClientsVersionHistoryModel from 'vault/models/clients/version-history';
 import type {
+  ByMonthClients,
   ByMonthNewClients,
+  MountByKey,
   MountNewClients,
   NamespaceByKey,
   NamespaceNewClients,
+  TotalClients,
 } from 'core/utils/client-count-utils';
 
 interface Args {
@@ -28,8 +31,11 @@ interface Args {
   endTimestamp: string;
   namespace: string;
   mountPath: string;
+  filteredByMonth: ByMonthClients[] | MountByKey[];
+  filteredTotals: TotalClients | undefined;
 }
 
+// Component class extended by Clients::Page::* components
 export default class ClientsActivityComponent extends Component<Args> {
   average = (
     data:
@@ -41,45 +47,11 @@ export default class ClientsActivityComponent extends Component<Args> {
   };
 
   get byMonthActivityData() {
-    const { activity, namespace } = this.args;
-    return namespace ? this.filteredActivityByMonth : activity.byMonth;
+    return this.args.filteredByMonth;
   }
 
   get byMonthNewClients() {
     return this.byMonthActivityData ? this.byMonthActivityData?.map((m) => m?.new_clients) : [];
-  }
-
-  get filteredActivityByMonth() {
-    const { namespace, mountPath, activity } = this.args;
-    if (!namespace && !mountPath) {
-      return activity.byMonth;
-    }
-    const namespaceData = activity.byMonth
-      ?.map((m) => m.namespaces_by_key[namespace])
-      .filter((d) => d !== undefined);
-
-    if (!mountPath) {
-      return namespaceData || [];
-    }
-
-    const mountData = namespaceData
-      ?.map((namespace) => namespace?.mounts_by_key[mountPath])
-      .filter((d) => d !== undefined);
-
-    return mountData || [];
-  }
-
-  get filteredActivityByNamespace() {
-    const { namespace, activity } = this.args;
-    return activity.byNamespace.find((ns) => ns.label === namespace);
-  }
-
-  get filteredActivityByAuthMount() {
-    return this.filteredActivityByNamespace?.mounts?.find((mount) => mount.label === this.args.mountPath);
-  }
-
-  get filteredActivity() {
-    return this.args.mountPath ? this.filteredActivityByAuthMount : this.filteredActivityByNamespace;
   }
 
   get isCurrentMonth() {
@@ -98,10 +70,9 @@ export default class ClientsActivityComponent extends Component<Args> {
     );
   }
 
-  // (object) top level TOTAL client counts for given date range
+  // (object) TOTAL client counts for given date range (filtered)
   get totalUsageCounts() {
-    const { namespace, activity } = this.args;
-    return namespace ? this.filteredActivity : activity.total;
+    return this.args.filteredTotals;
   }
 
   get upgradesDuringActivity() {
@@ -117,44 +88,5 @@ export default class ClientsActivityComponent extends Component<Args> {
     }
 
     return this.byMonthActivityData[0]?.new_clients;
-  }
-
-  // total client data for horizontal bar chart in attribution component
-  get totalClientAttribution() {
-    const { namespace, activity } = this.args;
-    if (namespace) {
-      return this.filteredActivityByNamespace?.mounts || null;
-    } else {
-      return activity.byNamespace || null;
-    }
-  }
-
-  // new client data for horizontal bar chart
-  get newClientAttribution() {
-    // new client attribution only available in a single, historical month (not a date range or current month)
-    if (this.isDateRange || this.isCurrentMonth || !this.newClientCounts) return null;
-
-    const newCounts = this.newClientCounts;
-    if (this.args.namespace && hasMountsKey(newCounts)) return newCounts?.mounts;
-
-    if (hasNamespacesKey(newCounts)) return newCounts?.namespaces;
-
-    return null;
-  }
-
-  get hasAttributionData() {
-    const { mountPath, namespace } = this.args;
-    if (!mountPath) {
-      if (namespace) {
-        const mounts = this.filteredActivityByNamespace?.mounts?.map((mount) => ({
-          id: mount.label,
-          name: mount.label,
-        }));
-        return mounts && mounts.length > 0;
-      }
-      return !!this.totalClientAttribution && this.totalUsageCounts && this.totalUsageCounts.clients !== 0;
-    }
-
-    return false;
   }
 }
