+++ ui/app/components/clients/page/overview.hbs
@@ -5,10 +5,10 @@
 
 <Clients::RunningTotal
   @isSecretsSyncActivated={{this.flags.secretsSyncIsActivated}}
-  @byMonthActivityData={{this.byMonthActivityData}}
+  @byMonthActivityData={{@filteredByMonth}}
   @isHistoricalMonth={{and (not this.isCurrentMonth) (not this.isDateRange)}}
   @isCurrentMonth={{this.isCurrentMonth}}
-  @runningTotals={{this.totalUsageCounts}}
+  @runningTotals={{@filteredTotals}}
   @upgradeData={{this.upgradesDuringActivity}}
   @responseTimestamp={{@activity.responseTimestamp}}
   @mountPath={{@mountPath}}
