+++ ui/app/templates/vault/cluster/clients/counts/sync.hbs
@@ -10,4 +10,6 @@
   @endTimestamp={{this.model.endTimestamp}}
   @namespace={{this.countsController.ns}}
   @mountPath={{this.countsController.mountPath}}
+  @filteredTotals={{this.countsController.filteredActivityTotals}}
+  @filteredByMonth={{this.countsController.filteredByMonthActivity}}
 />
